2023-02-21 a reggie tack out. Improve user management logic,and complete user -end development.

+++ src/main/java/com/item/reggie/controller/CategoryController.java
@@ -5,9 +5,12 @@ import com.baomidou.mybatisplus.annotation.FieldFill;
 import com.baomidou.mybatisplus.annotation.TableField;
 import com.baomidou.mybatisplus.core.conditions.query.LambdaQueryWrapper;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
+import com.item.reggie.dto.DishDto;
 import com.item.reggie.entity.Category;
 import com.item.reggie.entity.Employee;
+import com.item.reggie.filter.CheckFilter;
 import com.item.reggie.service.CategoryService;
+import com.item.reggie.service.DishService;
 import lombok.extern.slf4j.Slf4j;
 import com.item.reggie.common.R;
 import org.apache.commons.lang.StringUtils;
@@ -28,6 +31,8 @@ public class CategoryController {
 
     @Autowired
     private CategoryService categoryService;
+    @Autowired
+    private DishService dishService;
 
 
     /**
@@ -36,10 +41,14 @@ public class CategoryController {
      * @return
      */
     @PostMapping
-    public R<String> save( @RequestBody Category category){
-        log.info("新增菜品");
-        categoryService.save(category);
-        return R.success("新增分类");
+    public R<String> save(HttpServletRequest httpServletRequest, @RequestBody Category category){
+        if(CheckFilter.checkemployee(httpServletRequest)) {
+            log.info("新增菜品");
+            categoryService.save(category);
+            return R.success("新增分类");
+        }
+        return R.error("NOTLOGIN");
+
 
     }
 
@@ -50,22 +59,26 @@ public class CategoryController {
      * @return
      */
     @GetMapping("/page")
-    public R<Page> page (int page, int pageSize){
-        log.info("菜品： page={},pageSize={}",page,pageSize);
+    public R<Page> page (HttpServletRequest httpServletRequest, int page, int pageSize){
+        if(CheckFilter.checkemployee(httpServletRequest)) {
+            log.info("菜品： page={},pageSize={}",page,pageSize);
 
-        //这是一个分页构造器
-        Page<Category> pageInfo = new Page<>(page,pageSize);
+            //这是一个分页构造器
+            Page<Category> pageInfo = new Page<>(page,pageSize);
 
-        //t条件构造器，做数据库查询用
-        LambdaQueryWrapper<Category> queryWrapper =new LambdaQueryWrapper<>();
+            //t条件构造器，做数据库查询用
+            LambdaQueryWrapper<Category> queryWrapper =new LambdaQueryWrapper<>();
 
-        //排序条件
-        queryWrapper.orderByAsc(Category::getSort);
+            //排序条件
+            queryWrapper.orderByAsc(Category::getSort);
 
-        //执行查询
-        categoryService.page(pageInfo,queryWrapper);
+            //执行查询
+            categoryService.page(pageInfo,queryWrapper);
 
-        return  R.success(pageInfo);
+            return  R.success(pageInfo);
+        }
+        return R.error("NOTLOGIN");
+
 
     }
 
@@ -75,10 +88,13 @@ public class CategoryController {
      * @return
      */
     @DeleteMapping
-    public R<String> delete(Long ids){
-        log.info("delete id={}",ids);
-        categoryService.remove(ids);
-        return R.success("分类信息删除成功！");
+    public R<String> delete(HttpServletRequest httpServletRequest,Long ids){
+        if(CheckFilter.checkemployee(httpServletRequest)) {
+            log.info("delete id={}", ids);
+            categoryService.remove(ids);
+            return R.success("分类信息删除成功！");
+        }
+        return R.error("NOTLOGIN");
     }
 
     /**
@@ -87,10 +103,13 @@ public class CategoryController {
      * @return
      */
     @PutMapping
-    public R<String> update(@RequestBody Category category){
-        log.info("修改分类信息：{}",category);
-        categoryService.updateById(category);
-        return R.success("修改信息成功");
+    public R<String> update(HttpServletRequest httpServletRequest, @RequestBody Category category){
+        if(CheckFilter.checkemployee(httpServletRequest)){
+            log.info("修改分类信息：{}",category);
+            categoryService.updateById(category);
+            return R.success("修改信息成功");
+        }
+      return R.error("NOTLOGIN");
     }
 
 
@@ -111,5 +130,7 @@ public class CategoryController {
 
         return R.success(list);
     }
+
+
 
 }
